Reduce duplicate tests
