+++ packages/wrangler/src/__tests__/helpers/mock-workers-subdomain.ts
@@ -1,0 +1,29 @@
+import { rest } from "msw";
+
+import { createFetchResult, msw } from "./msw";
+
+/** Create a mock handler for the request to get the account's subdomain. */
+export function mockSubDomainRequest(
+	subdomain = "test-sub-domain",
+	registeredWorkersDev = true
+) {
+	if (registeredWorkersDev) {
+		msw.use(
+			rest.get("*/accounts/:accountId/workers/subdomain", (req, res, ctx) => {
+				return res.once(ctx.json(createFetchResult({ subdomain })));
+			})
+		);
+	} else {
+		msw.use(
+			rest.get("*/accounts/:accountId/workers/subdomain", (req, res, ctx) => {
+				return res.once(
+					ctx.json(
+						createFetchResult(null, false, [
+							{ code: 10007, message: "haven't registered workers.dev" },
+						])
+					)
+				);
+			})
+		);
+	}
+}
